Finish building a dataset for merging spe and bulk
build_database_04.ipynb: fix the bug found in build_database_03.ipynb
and redo the workflow. This time I adopt the pilot_test's way of
calculating composite depth, simply summing up the section start depth
in the filename and section depth. In this case, I don't need to deal
with core SO264-54-2 separately. The section depth can be found in the
filename and in the conent. I prefer to read from the content because it
has less chance to be accidently mess up by human. P.S. both chances are
low actually.

Also, I adopt the bulk dataset from database_build_03.ipynb and merge it
with the spe dataset produced here. spe+bulk_dataset_20201008.csv is the
output. It's ready for the later ML applications.