Build model for CaCO3 (grid search)
build_models_01.ipynb: import the latest  dataset from
build_database_04.ipynb and normalize it. Grid search for the
combinations of NMF+SVR and PCA+SVR. NMF+SVR provides best result. There
are some issues overcome and written in it. The question why the Ridged
regression doing good job in train-test splits but not the whole dataset
is still no answer yet. The scree plot of PCA, the visualizations for
the grid-search results, and train-test splits' data distribution are
produced in dates 20201009-1010.

.gitignore: since I started to use cluster to do the computation, there
are some files and scripts generated or used.