Improve performance and export the model of CaCO3
build_model_01.ipynb: Accidently find the max_iter for NMF is not
enough, so I increase it to 2000. In the meantime, I increase the folds
of CV from 5 to 10 to enhance the robustness of CV score. The best
parameters from the grid search remain the same, but the CV score is
slightly reduced to 0.97 while the test score is slightly increased.
The model with the best parameters is build and exported as
caco3_nmf_svr_model_20201011.joblib.

.gitignore: add the folder, models.