Buld models for CaCO3, TOC and TC
The model building, exporting and result visualization (data
distribution, grid search scores and accuracy in test set) are
consistently done or updated to the three notebooks. The train-test
split is asigned a random_state (24).

The issue of low R-squared in TOC is solved in build_models_02.ipynb. We
use the residuals as the error measure because the values of TOC are too
low.

The models are: caco3_nmf+svr_model_20201012.joblib,
toc_nmf+svr_model_20201012.joblib and tc_nmf+svr_model_20201012.joblib.